Create Plugin Abstraction
Plugin is the way to extend the functionality adding support to other
documentation formats and languages.
